requirements.txt
sanic[ext]
chess
pyjwt
gitpython